synchronize upload methods of ThumbnailUploadStep and VideoUploadStep in new class ThrottledUploadStep
