+++ STFU.Lib.Youtube/STFU.Lib.Youtube.csproj
@@ -42,6 +42,7 @@
     <Compile Include="Upload\JobUploader.cs" />
     <Compile Include="Upload\Steps\ChangeVideoDetailsStep.cs" />
     <Compile Include="Upload\Steps\RetryingUploadStep.cs" />
+    <Compile Include="Upload\Steps\ThrottledUploadStep.cs" />
     <Compile Include="Upload\Steps\ThumbnailUploadStep.cs" />
     <Compile Include="Upload\Steps\AbstractUploadStep.cs" />
     <Compile Include="Upload\Steps\UploaderStateChangedEventArgs.cs" />
